Move MalwareScanner and attachments into distinct folders
