+++ srv/attachments/azure-blob-storage.js
@@ -2,12 +2,12 @@ const { BlobServiceClient } = require("@azure/storage-blob")
 const { AbortController } = require("abort-controller")
 const cds = require("@sap/cds")
 const LOG = cds.log("attachments")
-const utils = require("../lib/helper")
+const utils = require("../../lib/helper")
 const {
   MAX_FILE_SIZE,
   sizeInBytes,
   createSizeCheckHandler,
-} = require("../lib/helper")
+} = require("../../lib/helper")
 
 module.exports = class AzureAttachmentsService extends (
   require("./object-store")
