+++ srv/attachments/basic.js
@@ -5,7 +5,7 @@ const {
   computeHash,
   traverseEntity,
   buildBackAssocChain,
-} = require("../lib/helper")
+} = require("../../lib/helper")
 
 class AttachmentsService extends cds.Service {
   init() {
