+++ srv/attachments/gcp.js
@@ -1,12 +1,12 @@
 const { Storage } = require("@google-cloud/storage")
 const cds = require("@sap/cds")
 const LOG = cds.log("attachments")
-const utils = require("../lib/helper")
+const utils = require("../../lib/helper")
 const {
   MAX_FILE_SIZE,
   sizeInBytes,
   createSizeCheckHandler,
-} = require("../lib/helper")
+} = require("../../lib/helper")
 
 module.exports = class GoogleAttachmentsService extends (
   require("./object-store")
