+++ srv/malware-scanner/malwareScanner.js
