+++ tests/unit/malwareScanner.test.js
@@ -3,7 +3,7 @@ const cds = require("@sap/cds")
 const { join } = cds.utils.path
 cds.test(join(__dirname, "../incidents-app"))
 
-const MalwareScanner = require("../../srv/malwareScanner")
+const MalwareScanner = require("../../srv/malware-scanner/malwareScanner")
 
 let scanner
 let attachmentsSvc
